Fix GCL 64-bit pointer truncation: change value_p to intptr_t (Option 2)
The fundamental issue: GCL_GetNextValue stored script data pointers in
int *value_p (4 bytes), truncating 64-bit pointers. The circular ptr table
workaround was fragile — entries got overwritten before bind triggers fired.

Fix: change GCL_GetNextValue's value_p parameter from int* to intptr_t*.
Pointers into GCL script data (GCLCODE_STRING, GCLCODE_SCRIPT_DATA,
GCLCODE_PARAMETER) are now stored directly as full 64-bit values.

Updated all callers in libgcl_fix/ (basic.c, command.c, expr.c, variable.c,
parse.c) and game_script_fix.c to use intptr_t for value variables.

The gcl_ptr_table is kept for backward compatibility but pointer-type
GCL values no longer need it — they go directly into the intptr_t variable.

This fixes:
- GCL_Command_if block pointer truncation (crash on stage scripts)
- HZD bind GCL block execution (camera zone transitions)
- All GCL script data references that contain embedded pointers

+++ source/game/delay.c
@@ -7,7 +7,7 @@
 #include "game/jimctrl.h"
 
 typedef union GCL_EXEC_UNION {
-    int            proc_id;
+    intptr_t       proc_id;
     unsigned char *block_top;
 } GCL_EXEC_UNION;
 
@@ -128,7 +128,7 @@ static void Act(Work *work)
     GV_DestroyActor(&work->actor);
 }
 
-void *GM_DelayedExecCommand(int proc, GCL_ARGS *args, int time)
+void *GM_DelayedExecCommand(intptr_t proc, GCL_ARGS *args, int time)
 {
     unsigned short argc;
     Work          *work;
